refactor(reminders): split SnoozeService and AlarmManagement out of AlarmManagerService
AlarmManagerService had three responsibilities: it was a BroadcastReceiver for
snooze intents, it held the core review reminder scheduling logic, and it was
meant to be the centralized point for handling AlarmManager in AnkiDroid.

Move the snoozing into SnoozeService, which also takes over the snooze pending
intent construction that lived in NotificationService, and move the alarm
window constant and notification bootstrapping into a new AlarmManagement
utility. AlarmManagerService is left with review reminder scheduling only and is
no longer a BroadcastReceiver; catchAlarmManagerExceptions joins AlarmManagement
in a later commit.
