+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiDroidApp.kt
@@ -54,7 +54,6 @@ import com.ichi2.anki.observability.ChangeManager
 import com.ichi2.anki.preferences.SharedPreferencesProvider
 import com.ichi2.anki.servicelayer.DebugInfoService
 import com.ichi2.anki.servicelayer.ThrowableFilterService
-import com.ichi2.anki.services.AlarmManagerService
 import com.ichi2.anki.services.NotificationService
 import com.ichi2.anki.settings.Prefs
 import com.ichi2.anki.settings.PrefsRepository
@@ -62,6 +61,7 @@ import com.ichi2.anki.startup.ensureCollectionPathSet
 import com.ichi2.anki.startup.getDefaultAnkiDroidDirectory
 import com.ichi2.anki.ui.dialogs.ActivityAgnosticDialogs
 import com.ichi2.anki.widget.RECURRING_WIDGETS
+import com.ichi2.utils.AlarmManagement
 import com.ichi2.utils.ExceptionUtil
 import com.ichi2.utils.LanguageUtil
 import com.ichi2.utils.measureTime
@@ -324,10 +324,13 @@ open class AnkiDroidApp :
             setupNotificationChannels(applicationContext)
 
             val context = this.withAppLocale()
+            // In the future, if more notifications are added to AnkiDroid, AlarmManagement.scheduleAllNotifications
+            // should be extended to handle them, but since for now the only modernized notifications are review reminder
+            // notifications, we block this behind the review-reminder-specific feature flag.
             if (Prefs.newReviewRemindersEnabled) {
-                Timber.i("Setting review reminder notifications if they have not already been set")
+                Timber.i("Setting notifications if they have not already been set")
                 applicationScope.launch {
-                    AlarmManagerService.scheduleAllNotifications(context)
+                    AlarmManagement.scheduleAllNotifications(context)
                 }
             } else {
                 // Register for notifications
