+++ AnkiDroid/src/test/java/com/ichi2/anki/services/AlarmManagerServiceTest.kt
@@ -4,7 +4,6 @@
 package com.ichi2.anki.services
 
 import android.app.AlarmManager
-import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
@@ -23,6 +22,7 @@ import com.ichi2.anki.reviewreminders.ReviewRemindersDatabase
 import com.ichi2.anki.services.NotificationService.Companion.EXTRA_REVIEW_REMINDER_ID
 import com.ichi2.anki.services.NotificationService.Companion.EXTRA_REVIEW_REMINDER_SCOPE
 import com.ichi2.anki.utils.ext.getParcelableCompat
+import com.ichi2.utils.AlarmManagement
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
@@ -36,8 +36,6 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.Calendar
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.minutes
 
 @RunWith(AndroidJUnit4::class)
 class AlarmManagerServiceTest : RobolectricTest() {
@@ -62,7 +60,6 @@ class AlarmManagerServiceTest : RobolectricTest() {
 
     private lateinit var context: Context
     private lateinit var alarmManager: AlarmManager
-    private lateinit var notificationManager: NotificationManager
     private lateinit var reviewReminder: ReviewReminder
 
     @Before
@@ -70,9 +67,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
         super.setUp()
         context = mockk(relaxed = true)
         alarmManager = mockk(relaxed = true)
-        notificationManager = mockk(relaxed = true)
         every { context.getSystemService<AlarmManager>() } returns alarmManager
-        every { context.getSystemService<NotificationManager>() } returns notificationManager
 
         reviewReminder = ReviewReminder.createReviewReminder(time = ReviewReminderTime(20, 0))
         reviewReminder.latestNotifTime = mockTime.intTimeMS() // Ensure the reminder is ready to have its future instance scheduled
@@ -100,7 +95,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -120,7 +115,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -135,12 +130,16 @@ class AlarmManagerServiceTest : RobolectricTest() {
             set(Calendar.HOUR_OF_DAY, 12)
             add(Calendar.DAY_OF_YEAR, 1)
         }
-        AlarmManagerService.scheduleReviewReminderNotification(context, currentTimeReviewReminder, attemptImmediateNotification = true)
+        AlarmManagerService.scheduleReviewReminderNotification(
+            context,
+            currentTimeReviewReminder,
+            attemptImmediateNotification = true,
+        )
         verify {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -188,7 +187,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
     }
 
     @Test
-    fun `scheduleAllNotifications schedules and fires reminders for all enabled reminders in database`() =
+    fun `scheduleAllEnabledReviewReminderNotifications schedules and fires reminders for all enabled reminders in database`() =
         runTest {
             val did1 = addDeck("Deck1")
             val did2 = addDeck("Deck2")
@@ -212,7 +211,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
                 )
             reviewReminders.forEach { ReviewRemindersDatabase.insertReminder(it) }
 
-            AlarmManagerService.scheduleAllNotifications(context)
+            AlarmManagerService.scheduleAllEnabledReviewReminderNotifications(context)
 
             reviewReminders.forEach { reminder ->
                 val expectedSchedulingTime = mockTime.calendar().clone() as Calendar
@@ -230,7 +229,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
                     alarmManager.setWindow(
                         AlarmManager.RTC_WAKEUP,
                         expectedSchedulingTime.timeInMillis,
-                        AlarmManagerService.WINDOW_LENGTH_MS,
+                        AlarmManagement.WINDOW_LENGTH_MS,
                         any(),
                     )
                 }
@@ -248,59 +247,4 @@ class AlarmManagerServiceTest : RobolectricTest() {
                     }.toSet()
             assertThat(firedReminderIds, equalTo(expectedFiredReminderIds))
         }
-
-    @Test
-    fun `triggering schedules snoozed notification and cancels clicked notification`() =
-        runTest {
-            ReviewRemindersDatabase.insertReminder(reviewReminder)
-
-            attemptSnooze(reviewReminder, 5.minutes)
-
-            verifyNotifSnoozed(5.minutes)
-            verifyPastNotifCleared(reviewReminder)
-        }
-
-    @Test
-    fun `triggering only clears past notif if review reminder is not in database`() =
-        runTest {
-            attemptSnooze(reviewReminder, 5.minutes)
-
-            verifyNoNotifSnoozed()
-            verifyPastNotifCleared(reviewReminder)
-        }
-
-    private suspend fun attemptSnooze(
-        reviewReminder: ReviewReminder,
-        delay: Duration,
-    ) {
-        AlarmManagerService.handleSnoozeReviewReminder(
-            context,
-            reviewReminder.id,
-            reviewReminder.scope,
-            snoozeIntervalInMinutes = delay.inWholeMinutes.toInt(),
-        )
-    }
-
-    private fun verifyNotifSnoozed(delay: Duration) {
-        verify(exactly = 1) {
-            alarmManager.setWindow(
-                AlarmManager.RTC_WAKEUP,
-                mockTime.intTimeMS() + delay.inWholeMilliseconds,
-                AlarmManagerService.WINDOW_LENGTH_MS,
-                any(),
-            )
-        }
-    }
-
-    private fun verifyNoNotifSnoozed() {
-        verify(exactly = 0) {
-            alarmManager.setWindow(AlarmManager.RTC_WAKEUP, any(), any(), any())
-        }
-    }
-
-    private fun verifyPastNotifCleared(reviewReminder: ReviewReminder) {
-        verify(exactly = 1) {
-            notificationManager.cancel(NotificationService.REVIEW_REMINDER_NOTIFICATION_TAG, reviewReminder.id.value)
-        }
-    }
 }
